Resolves conflicts preserving master's v7 kernel API (kzalloc_objs).

Changes from ubuntu branch:
- patches 1-3: comment style fixes, unsigned int, regulator refactor
  (VIO no consumers, VSIO→dovdd/INT3477, no always_on)
- ov8858: remove vsio supply name, remove ACPI_PTR(), fix UBSAN in dev_info
- build.sh: timestamp-based rebuild (skip up-to-date modules)
- install.sh: always calls build.sh, requires sudo, loopback service
  installed as real user via SUDO_USER
- README: updated for v4 regulator mapping and install workflow

+++ patches/tps68470/tps68470.c
@@ -147,20 +147,21 @@ static int skl_int3472_tps68470_probe(struct i2c_client *client)
 	struct tps68470_clk_platform_data *clk_pdata;
 	struct mfd_cell *cells;
 	struct regmap *regmap;
-	int n_consumers;
+	unsigned int n_consumers;
 	int device_type;
 	int ret;
-	int i;
+	unsigned int i;
 
 	if (!adev)
 		return -ENODEV;
 
 	/*
-	 * Look up board data before building clock platform data.  On platforms
-	 * where a sensor's ACPI _DEP does not list the INT3472 device,
-	 * for_each_acpi_consumer_dev() misses that sensor and its clock consumer
-	 * entry is never registered.  Board data can supply a static consumer
-	 * list to use instead, bypassing the broken _DEP traversal.
+	 * Look up board data before building clock platform data.  On
+	 * platforms where a sensor's ACPI _DEP does not list the INT3472
+	 * device, for_each_acpi_consumer_dev() misses that sensor and its
+	 * clock consumer entry is never registered.  Board data can supply
+	 * a static consumer list to use instead, bypassing the broken _DEP
+	 * traversal.
 	 */
 	board_data = int3472_tps68470_get_board_data(dev_name(&client->dev));
 
@@ -172,7 +173,7 @@ static int skl_int3472_tps68470_probe(struct i2c_client *client)
 		if (!clk_pdata)
 			return -ENOMEM;
 		clk_pdata->n_consumers = board_data->n_clk_consumers;
-		for (i = 0; i < (int)board_data->n_clk_consumers; i++)
+		for (i = 0; i < board_data->n_clk_consumers; i++)
 			clk_pdata->consumers[i] = board_data->clk_consumers[i];
 		n_consumers = board_data->n_clk_consumers;
 	} else {
@@ -254,7 +255,7 @@ static int skl_int3472_tps68470_probe(struct i2c_client *client)
 static void skl_int3472_tps68470_remove(struct i2c_client *client)
 {
 	const struct int3472_tps68470_board_data *board_data;
-	int i;
+	unsigned int i;
 
 	board_data = int3472_tps68470_get_board_data(dev_name(&client->dev));
 	if (board_data) {
